refactor(thumbnail): move from gallery dir to streets
