+++ assets/scripts/streets/__tests__/thumbnail.test.js
